+++ lib/download_tv/grabbers/tpb.rb
@@ -4,42 +4,27 @@ module DownloadTV
   ##
   # ThePirateBay grabber
   class ThePirateBay < LinkGrabber
-    TRACKERS = %w[
-      udp://tracker.coppersurfer.tk:6969/announce
-      udp://tracker.openbittorrent.com:6969/announce
-      udp://tracker.opentrackr.org:1337
-      udp://movies.zsw.ca:6969/announce
-      udp://tracker.dler.org:6969/announce
-      udp://opentracker.i2p.rocks:6969/announce
-      udp://open.stealth.si:80/announce
-      udp://tracker.0x.tf:6969/announce
-    ]
+    def initialize(tpb_proxy = 'https://tpb30.ukpass.co/')
+      proxy = tpb_proxy.gsub(%r{/+$}, '')
 
-    def initialize
-      super("https://tpb36.ukpass.co/apibay/q.php?q=%s&cat=")
+      super("#{proxy}/search/%s/0/7/0")
     end
 
     def get_links(show)
       search = format(@url, show)
 
-      data = agent.get(search)
-      parsed = JSON.parse(data.body)
+      # Skip the header
+      data = agent.get(search).search('#searchResult tr').drop 1
 
-      raise NoTorrentsError if parsed.size == 1 && parsed.first['name'] == 'No results returned'
+      raise NoTorrentsError if data.empty?
 
-      parsed.map do |elem|
-        [elem['name'], build_magnet(elem['info_hash'], elem['name'])]
-      end
-    end
+      # Second cell of each row contains links and name
+      results = data.map { |d| d.search('td')[1] }
 
-    private
+      names = results.collect { |i| i.search('.detName').text.strip }
+      links = results.collect { |i| i.search('a')[1].attribute('href').text }
 
-    def build_magnet(torrent_hash, name)
-      "magnet:?xt=urn:btih:#{torrent_hash}&dn=#{CGI.escape(name)}#{trackers_params}"
-    end
-
-    def trackers_params
-      trackers_params ||= "&tr=#{TRACKERS.map { |tracker| CGI.escape(tracker) }.join('&tr=')}"
+      names.zip(links)
     end
   end
 end